requirements.txt
fastapi==0.95.2
uvicorn==0.23.2
bcrypt==4.0.1
pydantic==1.10.7
supabase-py==0.1.12
python-dotenv==1.0.0